src/requirements.txt
pandas
pycryptodome
cryptography